pyproject.toml
[tool.poetry]
name = "information_robot"
version = "0.1.0"
description = "information robot"
authors = ["Iydon Liang <liangiydon@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.*"
graia-application-mirai = "^0.19.1"
graia-scheduler = "^0.0.4"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
ipython = "^7.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
